[svn r23] minor changes.
BUT IT WORKS NOW!

So, you should be able to go into a directory and run python2.5 test_fido.py
and it'll watch for new outputs, and call the makePlots handler that's in
there.  Right now it's set to be MY choice of plots, but you can easily modify
it to be your own set.