[svn r41] Projections can now be weighted by any (derived) field now.  For example, DensitySquared is a good one to use.
My additions slow do the non-weighted projections by 25-50%, but I
didn't want to create messy code with many if-then-elses, depending on
being weighted.  I figured that any projection besides a regular
density one will be weighted, so this won't be that big of a deal...

If it will be, then I can make separate projection instances for
non-weighted and weighted cases, so the original speed will be
untouched.

New example usage:
    plot.addProj("Temperature", 1, weight="DensitySquared")