[svn r47] Some primitive support for the SWIG interface.  You can now call:
EnzoHierarchy.initializeEnzoInterface()

and it'll read in the parameter file, initialize the rate data, and set up the
global values.  Right now this is only somewhat useful.  (Note that this sets
up yt.enki.EnzoInterface.cvar , among other things.)  The more useful thing
is:

EnzoGrid.getEnzoGrid()

which will call ReadGrid() on that grid, and then stores it as EnzoGrid.eiGrid,
which has all of the methods for that grid.  The  data interface is still a bit
clunky, but, excitingly enough you can call things like
Grid.SolveHydroEquations (with a bit of work.)  I'm currently using this to
test my new grid::SolveCoupledChemistryCooling method.

NOTE!  This will *probably* give bizarre and unexpected results if you load
multiple EnzoHierarchy's and instantiate EnzoInterfaces for each one, due to
Enzo's global variables (which are all stored in yt.enki.EnzoInterface.cvar)
which tend to pollute namespace.  I am thinking of a way around this.

The medium-term goal of this project will be to generate or modify data in
python code, and write them back out by calling WriteAllData().  ReadAllData is
going to be implemented shortly.  I am foreseeing the possibility of writing a
python wrapper to ReadAllData that reads the entire grid hierarchy, but does
not read in grids except as needed, by feeding in a 0 to grid::ReadGrid's last
argument.

Anyway, this SWIG stuff is fun, so see me if you are interested in using this
to, for instance, generate initial conditions using python code instead of C++.