[svn r85] Fixed up namespace bugs...
ALSO!  Potentially important bug fixed in the EnzoSlice datatype.

Previously, the EnzoSlice.coords array was generated *incorrectly*.  It was
always the *plot*-centric x,y,z, not the *data*-centric x,y,z.  So column 1 was
always whatever was the x coordinate on the hippodraw plot, column 2 was the y
coordinate from HD, and column 3 was always always ALWAYS a constant value.

This is incorrect.

So I have fixed it.  .x,.y and .z are still HD-centric, but self.coords, which
is what everyone should use when doing derived fields, is fixed.

Also, I changed all the range to xrange for iteration.  Saves a small amount of
memory.

(And I did find out why it takes up so much memory after loaidng some
hierarchies...  and I am working on a fix...  unfortunately, it looks like it
will involve some forking action in Fido.  Standalone processes will be
unaffected, and will continue to take up memory.  But that's desirable, I
think.)