[svn r165] Moving to a new backend system.  I am including the README here:
>>>
Testing directory, designed to refactor raven into using different backends.

This may, eventually, lead to using things like Chaco and MayaVi2.  (Ultimately
we want to have support for 3D stuff, but, alas, that is a bit off for now.)

For now, however, it's going to be fairly barebones, with simple engines for
Matplotlib and Hippodraw.  To accomplish this task, we are going to do the
following:

    -- Move all engine-centric things into backends/
    -- Define a generalized set of plot-types, which we'll then use the engine
       to plot.  Note that when I say "plot-types" here I mean it in a more
       data-centric manner than the backends themselves would mean.
    -- Define a set of functions that are required to be implemented by each
       backend engine, even if that engine simply passes on a plot-call.


    Matt Turk, mturk@slac.stanford.edu, 11JUN07


Preliminary skeleton of engines:

(f = function
 v = variable
 c = class   )

 f   Initialize
 f   CleanUp
 v   engineVals
 c   RavenPlot[Engine?]
 c     VMPlot
 c       ProjectionPlot
 c       SlicePlot
 c     LinePlot
 c       ProfilePlot
 c       ACProfilePlot
 c     RegionPlot
 c       TwoPhasePlot   ! Won't work in MPL for now
 c       ThreePhasePlot ! Use scatter in MPL?  (Supply axes)
 c     ScatterPlot
 c     HistogramPlot
<<<

HippoDraw interface is mostly functional, and significantly
shorter.  So that's a good thing!  Additionally, the Matplotlib one is not
really functional yet, but it will be, soon.