[svn r167] Two major additions:
Better support for AnalyzeCluster.
Preliminary Amira export support.  Probably needs some tweaking, but it loads
(although on this remote connection I can't tell if it is actually working or
not.)  In the future I will probably pass file-handles back and forth rather
than file names, as it's sort of slow right now, and that could improve it.  I
forget why I didn't this time, but I think I had Good Reason.  Or at least, a
reason that felt Good at the time.