CHANGES!
Profiler now works pretty memory-efficiently.  I can profile the Light Cone in
a couple hours on my G5 without any memory issues.  It's slow.  This is to be
expected, though, and I can't quite figure out how to pool connections in
pytables without using up too much memory.  The parallelization should not be
difficult, using the new ...

GridCollection object, which accepts a list of grids and then uses those as its
data source.  This is the first step toward paralellization in a transparent
fashion.

dx, dy, dz are now proper fields, and they are memory-conservative.

clear_data on a grid is no longer completely-destroying.  It can recover!

test_lagos now passes all the tests.

This is probably ready for a merge.