[svn r574] Mostly fixing and adding docstrings.
One *major* change.  The 'center' returned by find_max in the hierarchy object
will now be at the *center* of a cell, not the edge.  This brings it in line
with what is expected elsewhere in the code, but it might cause minor changes
in display attributes.

Got rid of the _get_projection method on the grid.  It's not needed.