[svn r1046] I'm sick of keeping track of the callbacks, so we're going to use a 'self-registering plugin' pattern, as per Effbot (http://effbot.org/zone/metaclass-plugins.htm)
Callbacks will now auto-register, and we simply import all of the
auto-registered callbacks.  The "exec" statement is kind of ugly, but I think
it's better than modifying the local dictionary.

The callback interface will likely change, too, as we will probably move to
adding them all as methods to RavenPlot, rather than manually doing
add_callback with an instance.