Adding AMR kd-Tree framework and rendering, as well as a few more options in the rendering process.
This involves several pieces of yt, detailed below:

yt/utilities/amr_kdtree/amr_kdtree.py - Can be used to partition an
amr hierarchy into kd-Tree whose leaves make up a homogenized,
non-overlapping, volume.  Details and usage can be found in the
docstrings for the AMRKDTree class.  The main object is the
AMRKDTree.tree, which is a dictionary whose keys are unique binary
tree ids that point to a position in the kd-Tree.  The kd-Tree
framework allows for fast location of positions, and
viewpoint-independent ordering of homogenized bricks, ideal for the
volume rendering.

yt/visualization/volume_rendering/camera.py - This camera object has
been updated to, by default, use the kd-Tree for volume decomposition
and ray casting.  Basic usage of the camera object has remained the
same, though there are many new options that can be seen in the
docstrings for the Camera class.  An email accompanying this commit
will detail the advantages, options, and several examples.

yt/data_objects/grid_patch.py - Modified to accept a keyword in
get_vertex_centered_data(..., no_ghost=False).  When no_ghost=True,
instead of gathering the ghost zones from surrounding grids, the ghost
outer vertices are filled by extrapolation from within the grid.  This
leads to very large (up to 300x in some cases) speedup, but is not as
accurate.  Details of when this option should be used are in the
Camera class docs.

yt/visualization/volume_rendering/transfer_functions.py - Added a
keyword col_bounds=None that specifies the colorbar range in the
transfer function when using sample_colormap or add_layers.  This
confines the colormap to a range of values that can differ from the
transfer function bounds themselves.  It has made it easier to sample
a full colormap over a reduced set of values.