This adds a second derived field check after particle types have been created.
What this addresses is when a baryon field and a particle field are combined.
Currently the ordering of field construction makes one available and the other
not available.  See #601 for more details.

I would prefer to see the solution be to re-generate dependencies as needed,
but we are not yet able to do that.