Fix failing test by being explicit.
dobj["all", "Coordinates] is currently not working for load_particles()
datasets.  I am not entirely certain why, but it does work in the unitrefactor,
and I think the field systems have diverged enough that I do not want to patch
it here when we will transition to that at some point in the near future.