Considerably simplifying smoothing.
This finishes implementing the volume-weighted smoothing suggested by
James Wadsley.  The units come out correctly, and the values look reasonable.
However, there are a few lingering issues with getting NaNs in the resulting
output; I believe this is related to not traversing all of the octs when
computing the values; some octs may only be refined from one particle species,
rather than the particle species being examined.  I will be attempting to
address this problem shortly.