Be much more careful about sanitizing data object centers.  Closes #848
The selector objects expect data in code units, so we need to be careful to set everything as such.

input YTArrays may come from a completely different dataset, as is the case in the halo catalog
code, so we also need to make sure we conver to CGS if passed a YTArray or list of quantities.