Deal with the project state after a bugfix release.
We now get the lineage of commits from the last "major" release (e.g. yt 3.2.0)
and use that to find merged PRs. Before we were going back merely to the last
release which is incorrect.  We still disregard PRs that hapenned before the last
minor release.

One caveat is that we're now no longer dealing with commits that were pushed
directly to the main repo and thus didn't come in as part of a pull request.
Supporting this would make the script much slower (we'd need to look at *all*
of the pull requests that came in since the last major release, not just the
ones that came in since the last bugfix release) and more complicated. Since
it's a rare event anyway, I'm just going to go ahead and punt on this.