Removing boolean data objects
These are currently untested, not being used, and would take a lot of work
to get working again. Rather than leaving the dead code in the codebase I am
removing it in the interest of code clarity. If someone wants to bring it back
they can use the VCS history to restore the code.