Removing tests in the top-level tests folder and old answer testing support
These tests are not run by the current testing infrastructure. In the interest
of clarity to new users (who might think these tests are useful since they are
at the root of the repository) I'm removing them. I've left in the test enzo
dataset to avoid adding 2.4 MB to the repository.