Fix some types of octree block traversal.
When either empty block is encountered or when the block is of shape (1,1,1),
special case the traversal such that the buffers are of the right shape.  This
can show up most frequently in ARTIO simulations where the root grid is
fully-refined, or where data objects touch root grid objects but only the
refined ones.  Or some other cases I'm not 100% sure of.