Allow adding or subtracting arrays filled with zeros without checking units
This is a behavior change that allows the python builtin sum() function to work
correctly. It also matches the behavior of astropy's units.

Overall this should have minimal performance implications for scripts that did
not raise errors before this pull request, since most code changes are in paths
that would have raised errors before.

There is one place where we need to see if there are any nonzero elements in
either array to ensure that units are propogated properly. Hopefully for most
real data that isn't filled with zeros this will be a relatively fast, but it
is a bit of overhead. Any suggestions on how to avoid doing it would be very
much appreciated.

I've also added tests for this new behavior, and added a test for the `sum`
builtin when applied to a YTArray.