Refactor constructors
All the common parts in related constructors are gathered together and put in the proper class. For `TipsyDataset`, `GadgetDataset` and `HTTPStreamDataset`, the API remains the same. For `ParticleDataset` and `SPHDataset` the API is expanded a little, which is still compatible with the original API. The reason for the expansion is that I want to keep the API consistent with `Dataset`.