Refactor implementation of the default kernel name functionality
First, their is a class attribute `default_kernel_name`. During instantiation, if the `kernel_name` argument is not set, the `default_kernel_name` will be used for the instance's `kernel_name` attribute. Then for the instance's `add_smoothed_particle_field` method, if the `kernel_name` argument is not set, the instance's `kernel_name` attribute will be used.